refactor: rename components for consistency
