+++ src/components/ui/FavoritesIcon.tsx
