+++ src/containers/UserSearchContainer.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent, FC } from 'react'
 import { Container, Spinner } from 'react-bootstrap'
-import AlertComponent from '../components/AlertComponent'
-import { PaginationComponent } from '../components/PaginationComponent'
 import Search from '../components/Search'
+import AlertComponent from '../components/ui/AlertComponent'
+import { PaginationComponent } from '../components/ui/PaginationComponent'
 import { UserSearchList } from '../components/UserSearchList'
 import { MOBILE_BREAKPOINT } from '../constants'
 import { useUserFavorites } from '../context/userFavoritesContext'
